Adds sketch for _submit_spark, drafts for run and execNode, skeletons for execMapNode, execJoinNode, execIterables
In order to get started here, I've put together a draft for whta I
think run and execNode should do, as well as a sketch for how
_submit_spark will run and function definitions for the other
necessary functions (though they don't contain anything).

run: takes flatgraph, execgraph, config, and updatehash from workflow,
sets _config and creates an OrderedDict where keys are the flat graph
nodes (in topological order) and values are the associated py
files. Those pyfiles are created at this time (as are the associated
pkl files for the nodes). This OrderedDict and the execgraph are then
passed to _submit_spark.

_submit_spark: this doesn't work as is, but I wanted to get ideas
down. It will need to parallelize the initial pyfiles and broadcast
the execgraph. It then iterate through the nodes in the OrderedDict,
making sure the RDD contents are what we assume them to me (the
corresponding pyfiles) and perform the equivalent operation (based on
the type of node).

execNode: this simply lodas in the pyfile so we can find the node,
exec the pyfile, and return the successors of the node, using the
broadcasted execgraph.